requirements.txt
langchain==0.0.296
openai==0.27.10
tiktoken==0.4.0
faiss-cpu==1.7.4
python-dotenv==1.0.0